demo/rag-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
numpy==1.26.4
opentelemetry-sdk==1.27.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
sse-starlette==2.1.0
